feture/ Se agrega cambios para obtener disponibilidad de reservas

+++ src/main/java/com/ludoteca/api/controller/TurnoController.java
@@ -72,7 +72,7 @@ public class TurnoController {
     // Eliminar turno día (ADMIN)
     @PreAuthorize("hasRole('ADMIN')")
     @DeleteMapping("/dias/{id}")
-    public ResponseEntity<Void> eliminarTurnoDia(@PathVariable final Long id) {
+    public ResponseEntity<Void> eliminarTurnoDia(@PathVariable final Integer id) {
         turnoDiaService.eliminar(id);
         return ResponseEntity.noContent().build();
     }
